pyproject.toml
[project]
name = "kasparro-backend-etl"
version = "0.1.0"
description = "Production-grade Backend & ETL system"
requires-python = ">=3.12"

dependencies = [
  "fastapi>=0.111.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.7.0",
  "sqlalchemy>=2.0.30",
  "asyncpg>=0.29.0",
  "python-dotenv>=1.0.1",
  "httpx>=0.27.0",
  "tenacity>=8.3.0",
  "pytest>=9.0.2",
  "pytest-asyncio>=1.3.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.6",
  "testcontainers>=4.7.0",
  "ruff>=0.4.5",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
